pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "placeholder_name"
dynamic = ["version"]
description = "Code Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.0.0",
    "httpx>=0.28.1",
    "hydra-core>=1.3.2",
    "icontract>=2.6.0",
    "lagom>=2.0.0",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.0.0",
    "returns>=0.26.0",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "transitions>=0.9.0",
    "typer>=0.20.0",
    "uvloop>=0.22.1",
]

[tool.uv]
package = true
dev-dependencies = [
    "ruff",
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "pytest-xdist",
    "pytest-cov",
    "ty",
    "vulture",
    "pip-audit",
    "hypothesis>=6.0.0",
    "anyio[trio]>=4.0.0",
    "detect-secrets>=1.5.0",
    "import-linter>=2.0",
    "bandit[toml]>=1.7.0",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "RUF", "ANN401", "PLC0415"]
fixable = ["ALL"]
# S101: allow assert in tests
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.vulture]
min_confidence = 80
paths = ["src", "main.py"]

[tool.ruff.format]
quote-style = "double"

[tool.hatch.version]
source = "vcs"

[tool.ty]
environment = { python-version = "3.12" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow — excluded from pre-commit fast-tests (deselect with '-m not slow')",
]
addopts = [
    "-n", "auto",
    "--cov=src",
    "--cov-report=term",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/.*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.importlinter]
root_packages = ["placeholder_name"]

[[tool.importlinter.contracts]]
name = "Core must not import from adapters"
type = "forbidden"
source_modules = ["placeholder_name.core"]
forbidden_modules = ["placeholder_name.adapters"]
